// lib/runtime/src/types/node-input.ts
import type { NodeId, SlotName } from '../pipeline-compat'

interface BaseNodeInput {
  id: NodeId
  slot?: SlotName
  children?: ReadonlyArray<NodeInput>
}

export interface NativeNodeInput extends BaseNodeInput {
  kind: 'native'
  tag: string
}

export interface ComponentNodeInput extends BaseNodeInput {
  kind: 'component'
}

export type NodeInput = NativeNodeInput | ComponentNodeInput